move AnnotationsPanel back
