+++ vcell-client/src/main/java/cbit/vcell/client/desktop/biomodel/BioModelEditorModelPanel.java
@@ -61,7 +61,6 @@ import javax.swing.event.DocumentListener;
 import javax.swing.event.ListSelectionEvent;
 import javax.swing.event.ListSelectionListener;
 
-import cbit.vcell.client.desktop.biomodel.annotations.AnnotationsPanel;
 import org.vcell.model.rbm.MolecularType;
 import org.vcell.model.rbm.RbmNetworkGenerator.CompartmentMode;
 import org.vcell.model.rbm.RbmUtils;
